move StoreProvider to main.jsx to wrap all components
