+++ src/App.jsx
@@ -2,18 +2,15 @@
 import './App.css';
 import { Outlet } from "react-router";
 import Nav from "./components/Navbar";
-import StoreProvider from "./redux/GlobalState";
 
 
 function App() {
   return (
     <>
-      <StoreProvider>
-        <Nav />
-        <main>
-          <Outlet />
-        </main>
-      </StoreProvider>
+      <Nav />
+      <main>
+        <Outlet />
+      </main>
     </>
   )
 }
