+++ src/redux/reducers/userReducer.js
@@ -11,18 +11,19 @@ export const USER_ACTIONS = {
 export const userReducer = (state = initialState, action) => {
   switch (action.type) {
     case USER_ACTIONS.LOGIN:
-      
+      console.log("@user login");
       return {
         ...state,
       }
 
     case USER_ACTIONS.SIGNUP:
-      
+      console.log("@user signup");
       return {
         ...state,
       }
 
     case USER_ACTIONS.LOGOUT:
+      console.log("@user logout");
       auth.signOut();
       return {
         ...state,
