+++ src/redux/store.js
@@ -1,4 +1,4 @@
-import { reducer } from "./reducers/reducers";
+import { reducer } from "./reducers/index.js";
 import { configureStore } from "@reduxjs/toolkit"
 
 export default configureStore({
